certyfikat
poprawki wprzechodzeniu do certyfikatu z listy certyfikatow uzytkownikow (admin) oraz poprwka deaktywacji oraz przedluzenia certyfikatow z pozycji admina

+++ Aplikacja_Mobilna/App_IK/app/src/main/java/inteligenty_zamek/app_ik/inWork/certyficatPresenter.kt
@@ -1,8 +1,11 @@
 package inteligenty_zamek.app_ik.inWork
 
 import android.content.Intent
+import android.util.Log
 import inteligenty_zamek.app_ik.API.HTTPRequestAPI
 import inteligenty_zamek.app_ik.API.fileReadWriteApi
+import inteligenty_zamek.app_ik.Views.GenerationCertyficatActivity
+import inteligenty_zamek.app_ik.Views.ManagmentCertyficationUserActivity
 import inteligenty_zamek.app_ik.Views.Managment_certyficationActivity
 import inteligenty_zamek.app_ik.rest_class.GlobalContainer
 import org.json.JSONArray
@@ -67,9 +70,24 @@ class certyficatPresenter(val view: certyficatActivity)
             val toSend: HashMap<String, String> = HashMap()
             toSend.put("login", model.login)
             toSend.put("token", model.token)
-            toSend.put("certificate_id",model.certyficat!!.idKey)
+            if(model.fromAdminPanel)
+            {
+                toSend.put("user_id",model.certyficat!!.id_user)
+            }
+            else
+            {
+                toSend.put("certificate_id",model.certyficat!!.idKey)
+            }
             try {
-                HTTPRequestAPI(this, "http://" + model.ipaddres + ":8080/api/deactivation/", "deleteResult", toSend).execute()
+
+              if(model.fromAdminPanel)
+              {
+                  HTTPRequestAPI(this, "http://" + model.ipaddres + ":8080/api/admin/deactivation_user_certificate/", "deleteResult", toSend).execute()
+              }
+                else
+              {
+                  HTTPRequestAPI(this, "http://" + model.ipaddres + ":8080/api/deactivation/", "deleteResult", toSend).execute()
+              }
             } catch (e: Exception) { }
 
 
@@ -82,12 +100,21 @@ class certyficatPresenter(val view: certyficatActivity)
 
   fun deleteResult(response:String)
   {
+      Log.i("yyyy",response)
       try {
           if ( JSONObject(response).getString("status") == "ok")
           {
-              deleteCertyficatFromFile()
-              val myIntent = Intent(view, Managment_certyficationActivity::class.java)
-              view.startActivity(myIntent)
+              if(model.fromAdminPanel)
+              {
+                  val myIntent = Intent(view, ManagmentCertyficationUserActivity::class.java)
+                  view.startActivity(myIntent)
+                  view.finish()
+              }
+              else {
+                  deleteCertyficatFromFile()
+                  val myIntent = Intent(view, Managment_certyficationActivity::class.java)
+                  view.startActivity(myIntent)
+              }
           }
           else
           {
@@ -98,15 +125,25 @@ class certyficatPresenter(val view: certyficatActivity)
 
     fun extendCertyficat()
     {
+        if(model.fromAdminPanel)
+        {
+            GlobalContainer.obj=model.certyficat
+            val myIntent = Intent(view, GenerationCertyficatActivity::class.java)
+            myIntent.putExtra("haveCertificate",true)
+            view.startActivity(myIntent)
+            view.finish()
+        }
+        else {
+            val toSend: HashMap<String, String> = HashMap()
+            toSend.put("login", model.login)
+            toSend.put("token", model.token)
+            toSend.put("lock_id", model.certyficat!!.id_lock)
+            try {
+                HTTPRequestAPI(this, "http://" + model.ipaddres + ":8080/api/request_new_certificate/", "extendResult", toSend).execute()
+            } catch (e: Exception) {
 
-        val toSend: HashMap<String, String> = HashMap()
-        toSend.put("login", model.login)
-        toSend.put("token", model.token)
-        toSend.put("lock_id",model.certyficat!!.id_lock)
-        try {
-            HTTPRequestAPI(this, "http://" + model.ipaddres + ":8080/api/request_new_certificate/", "extendResult", toSend).execute()
-        } catch (e: Exception) { }
-
+            }
+        }
     }
 
     fun extendResult(response:String) {
